closes #3735 and #4485
from 27s down to 5s, together with #4484
and 13s down to 5s for this PR alone
```r
library(data.table)
allIterations <- data.frame(v1 = runif(1e5), v2 = runif(1e5))
DoSomething <- function(row) someCalculation <- row[["v1"]] + 1
system.time(for (r in 1:nrow(allIterations)) DoSomething(allIterations[r, ]))
#   user  system elapsed 
#  3.384   0.007   3.392
allIterations <- as.data.table(allIterations)
setDTthreads(1)
system.time(for (r in 1:nrow(allIterations)) DoSomething(allIterations[r, , with=c(i=FALSE)]))
#   user  system elapsed 
#  5.432   0.121   5.554 
```
Once the general idea and API will be approved, the following items should be added:
- extra escapes of `with` and `j` processing in downstream code in `[.data.table`
- manual
- more tests
- news