@MichaelChirico 2020-05-25
basic API looks good. Do you know if `with=FALSE` now is faster than an equivalent `with=TRUE` query? I worry if we build this as "faster row selection` and `with=FALSE` is _not_ "faster column selection" it would be confusing