@jangorecki 2020-05-25
> Do you know if with=FALSE now is faster than an equivalent with=TRUE query?

if it is `identical(with, FALSE)` then new optimization is totally escaped, to stay backward compatible, so is not faster. `with=c(j=FALSE)` will be faster, but won't be replacement of `with=FALSE` because it will not handle `!"a"` or `a:b`. It will be really non-NSE.

> I worry if we build this as "faster row selection" and with=FALSE` is not "faster column selection" it would be confusing.

The "faster" is not that much relevant here, I would call it low overhead, which eventually can be faster when looping many times. Otherwise speed difference is insignificant.