@MichaelChirico 2020-05-26
OK. Still if "low-overhead-ness" differs between `with=FALSE` and `with = c(i=FALSE)`, it may be confusing.