@jangorecki 2020-05-26
`with=FALSE` is not really low-overhead, `with=c(j=FALSE)` is. Agree it deserves good documentation. The good thing is that change is fully backward compatible, so use of `with=FALSE` won't be affected.