@ColeMiller1 2020-05-26
Most options are lists, including discussion of ```nomatch = NULL``` discusses the use of a list. For consistency, I wonder if ```with = .(i = TRUE, j = FALSE)``` would be more ```data.table```ish. 

For some parts, maybe we could include ``` `.` = function(...) list(...)``` so we do not have to do NSE.