@jangorecki 2020-05-26
@ColeMiller1 to make this NSE (so the argument can be evaluated in separate line and result passed to `with`) we would need to export `.` which is not a good idea. Such "consistency" doesn't make much sense, if we go that way we will ended up using `allow.cartesian=.(TRUE)` as well. Let's use that when it can solve a new problem or overcome existing limitation.