@ColeMiller1 2020-05-30
Is using the ```[.data.frame``` subset method an option? For small subsets of ```i```, ```[.data.frame``` is faster than ```CsubsetDT```. On my Windows 10 with 1 thread, I see (timings all also include ```doSomething()``` wrapped around subset):

|Timing|Code|
|-------|-------|
| 5.1 s |```df[r, ]```|
| 5.3 s |``` `[.data.frame`(dt, r,)```|
|17.3 s| ```dt[r, , with = c(i = FALSE))]```|
|26.7 s| ```dt[r, ]```|

<details>
  <summary>Code run to make benchmarks</summary>

``` r
library(data.table)
setDTthreads(1L)

DoSomething <- function(row) someCalculation <- row[["v1"]] + 1
allIterations <- data.table(v1 = runif(1e5), v2 = runif(1e5))

system.time(for (r in 1:nrow(allIterations)) DoSomething(allIterations[r, ]))
#>    user  system elapsed 
#>   21.01    4.58   26.74
system.time(for (r in 1:nrow(allIterations)) DoSomething(`[.data.frame`(allIterations, r,)))
#>    user  system elapsed 
#>    5.09    0.00    5.27
identical(DoSomething(allIterations[r, ]), DoSomething(`[.data.frame`(allIterations, r,)))
#> [1] TRUE
all.equal(allIterations[r, ], `[.data.frame`(allIterations, r,))
#> [1] TRUE

#############Using this branch############
##remotes::install_github("Rdatatable/data.table", ref = "with2")
system.time(for (r in 1:nrow(allIterations)) DoSomething(allIterations[r, , with=c(i=FALSE)]))

#>    user  system elapsed 
#>   12.33    4.80   17.86

#############DF only##############
setDF(allIterations)
system.time(for (r in 1:nrow(allIterations)) DoSomething(allIterations[r, ]))
#>    user  system elapsed 
#>    4.97    0.00    5.08
```
</details>